Migrate victory-candlestick to TypeScript
