+++ .changeset/nine-cups-complain.md
@@ -1,0 +1,5 @@
+---
+"victory-candlestick": patch
+---
+
+Migrate victory-candlestick to TypeScript
