+++ packages/victory-candlestick/src/helper-methods.test.ts
