--- packages/victory-candlestick/src/index.js
@@ -1,2 +1,0 @@
-export { default as VictoryCandlestick } from "./victory-candlestick";
-export { default as Candle } from "./candle";
