+++ packages/victory-candlestick/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./victory-candlestick";
+export * from "./candle";
